Refactored bufferManagerImpl to allow an option in the constructor to use a hard reference early buffer. Modify the unit test as appropriate.
