+++ dev/com.ibm.ws.logging_test/bnd.bnd
@@ -52,6 +52,7 @@ test.project: true
 	org.jmock:jmock-legacy;version=2.5.0, \
 	com.ibm.ws.org.objenesis:objenesis;version=1.0, \
 	org.eclipse.osgi;version=latest, \
+    com.ibm.ws.logging.core;version=latest,\
     com.ibm.ws.logging;version=latest,\
     com.ibm.ws.org.apache.yoko.corba.spec.1.5;version=latest,\
     com.ibm.ws.org.apache.yoko.core.1.5;version=latest,\
