+++ dev/com.ibm.ws.logging_test/test/com/ibm/ws/collector/manager/buffer/SimpleRotatingQueueTest.java
@@ -8,13 +8,15 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.logging.internal.impl;
+package com.ibm.ws.collector.manager.buffer;
 
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertSame;
 
 import org.junit.Test;
+
+import com.ibm.ws.logging.internal.impl.SimpleRotatingQueue;
 
 /**
  *
